requirements.txt
Flask==1.1.2
gunicorn==20.0.4
werkzeug==1.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.3
Flask-Login==0.5.0
flask_socketio==5.0.1
requests==2.23.0
wtforms==2.3.1
flask_wtf==0.14.3
psycopg2==2.8.6
Alembic==1.4.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
eventlet==0.30.2
pyjwt==1.4.2

